+++ include/types.h
@@ -12,13 +12,13 @@
 
 // CNV candidate location map
 // (chr, snp_pos) : cnv_type
-using SNPLocation = std::pair<std::string, int>;
+using SNPLocation = std::pair<std::string, int64_t>;
 using SNPToCNVMap = std::map<SNPLocation, int>;
 
 // SV candidate read depth map. An SV is defined by its location, type, and
 // alternate allele.
 // (chr, start, end, sv_type, alt_allele) : (ref_allele, num_reads)
-using SVCandidate = std::tuple<std::string, int, int, std::string>;  // chr, start, end, alt_allele
+using SVCandidate = std::tuple<std::string, int64_t, int64_t, std::string>;  // chr, start, end, alt_allele
 // using SVInfo = std::tuple<int, int, std::string>;  // SV type, read depth, alignment type (CIGAR or split read)
 
 // // Create a class for storing SV information
@@ -46,7 +46,7 @@ using SVDepthMap = std::map<SVCandidate, SVInfo>;  // Map for getting type and r
 
 // SV calling:
 // Alignment location (chr, start, end, depth)
-using AlignmentData = std::tuple<std::string, int, int, int>;
+using AlignmentData = std::tuple<std::string, int64_t, int64_t, int>;
 using AlignmentVector = std::vector<AlignmentData>;
 
 // Query map (query name, alignment vector)
